Cryptography/requirements.txt
matplotlib
numpy
Pillow
tqdm